rename files: namespace.rs -> payload_prover.rs
